refactor(ModelDict): move to models
